% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package attach hook for STATSports API credentials}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The name of the package library (automatically provided)}

\item{pkgname}{The name of the package (automatically provided)}
}
\description{
`.onAttach()` displays credential load status messages to the user.
}
\keyword{internal}
