% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imu.R
\name{get_imu_only}
\alias{get_imu_only}
\title{Get STATSports IMU Data}
\usage{
get_imu_only(rawDataId)
}
\arguments{
\item{rawDataId}{Raw Data ID of a Players GPS Unit (Recommended; run \code{get_session()} to get available rawDataId's of intended session)}
}
\value{
A data frame containing Raw IMU GPS Data
}
\description{
Queries the getSessionImuData API Endpoint to retrieve only Raw GPS IMU data
}
