% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get All Available STATSports GPS Metrics}
\usage{
get_metrics()
}
\value{
A data frame containing ALL GPS Metrics.
}
\description{
Queries the GetAvailableMetrics API Endpoint to retrieve all available GPS Metrics.
}
