% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get_profiles}
\alias{get_profiles}
\title{Get STATSports Athlete profiles}
\usage{
get_profiles()
}
\value{
A data frame containing STATSports athlete profile information for the stored `api_key`.
If no profiles are found, the function raises an error.
}
\description{
Queries the GetPlayerDetails API Endpoint and returns a data frame of athlete profiles.
}
