% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw.R
\name{get_raw}
\alias{get_raw}
\title{Get STATSports Raw Data}
\usage{
get_raw(rawDataId)
}
\arguments{
\item{rawDataId}{Raw Data ID of a Players GPS Unit (Recommended; run \code{get_session()} to get available rawDataId's of intended session)}
}
\value{
A data frame containing Raw IMU GPS Data
}
\description{
Queries the getSessionRawData API Endpoint to retrieve Raw GPS data
}
