% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session.R
\name{get_session}
\alias{get_session}
\title{Get STATSports Session(s) and Drills}
\usage{
get_session(session_date, drills = TRUE, metrics = NULL)
}
\arguments{
\item{session_date}{ISO 8601 UTC date string (e.g., "2025-06-25").}

\item{drills}{Logical TRUE or FALSE to include drill breakdown.}

\item{metrics}{Optional character vector of available metrics to filter by (Recommended; run \code{get_metrics()} to get all available metrics)}
}
\value{
A data frame containing GPS Session results matching the optional filters.
}
\description{
Queries the GetFullSession API Endpoint to retrieve GPS Session and Drill data
}
